Creating Authorization Policies and pagination done

+++ app/Http/Controllers/Admin/AdminPostController.php
@@ -4,25 +4,28 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Post;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
-use Illuminate\Support\Facades\Storage;
 
 class AdminPostController extends Controller
 {
 
     public function index()
     {
-        $posts = Post::all();
+        $posts = Post::latest()->paginate(3);
+//        $posts = auth()->user()->posts;
         return view('admin.posts.index-post',['posts'=>$posts]);
     }
     public function create()
     {
+        $this->authorize('create',Post::class);
+
         return view('admin.posts.create-post');
     }
 
     public function store()
     {
+        $this->authorize('create',Post::class);
+
         $data = \request()->validate([
             'title'=>'required|min:8|max:255',
             'post_image'=>'file',
@@ -42,6 +45,8 @@ class AdminPostController extends Controller
 
     public function edit(Post $postsAdmin)
     {
+        $this->authorize('update',$postsAdmin);
+
         return view('admin.posts.edit-post',['postsAdmin'=>$postsAdmin]);
     }
 
@@ -61,6 +66,8 @@ class AdminPostController extends Controller
             $postsAdmin->post_image = $data['post_image'];
         }
 
+        $this->authorize('update',$postsAdmin);
+
         auth()->user()->posts()->save($postsAdmin);
 
         Session::flash('post-updated','Post with title: "' .$data['title']. '" has been updated');
@@ -70,6 +77,8 @@ class AdminPostController extends Controller
     }
     public function destroy(Post $postsAdmin)
     {
+        $this->authorize('delete',$postsAdmin);
+
         $postsAdmin->delete();
 
         Session::flash('post-deleted','Post with title: "' .$postsAdmin->title. '" has been deleted');
